giant data set + refined augmentation

+++ combination_with_mask.py
@@ -11,7 +11,7 @@ import os
 from math import degrees
 
 
-mode = "pole"
+mode = "test"
 
 write_path = "/Users/will/projects/legoproj/data_oneofeach/{}_oneofeach/".format(mode)
 
@@ -100,7 +100,7 @@ def shadeMasks(objects, x, objdata):
             scene.render.resolution_percentage = 100
                     
             scene.render.image_settings.file_format = 'PNG'
-            scene.render.filepath = write_path + str(x) + "_" + mode + "_" + obj.name + ".png"
+            scene.render.filepath = write_path + str(x) + "_" + mode + "_" + obj.name.replace(".","_") + ".png"
             bpy.ops.render.render(write_still = 1)
             count+=1
 
@@ -127,18 +127,21 @@ def genPiece(center):
         obj = objcopy(pole)
         mult = random.randint(-1,1)
         obj.location = addtups( center , tuple(mult * x for x in posm) )
+        
         pt = 90 if mult <= 0 else -90
         pt = pt + .7 * mult * 50
+        
         obj.rotation_euler = (0,0, math.radians(pt))
-        #print("Generating pole")
-        return 'Brick', obj
+
+        return 'Pole', obj
 
     else:
-        obj = objcopy(pole)
+        obj = objcopy(brick)
         pt = random.randint(0,20)/20
+        
         obj.rotation_euler = (0,0,pt * PI)
         obj.location = addtups( center , mltup(posm,.6) )
-        #print("Generating brick")
+
         return 'Brick', obj
 
 
@@ -147,14 +150,14 @@ def genWing(center):
 
     print("Generating wing")
     
-    '''if True or gimme() or gimme():
+    if True or gimme() or gimme():
         newWing = objcopy(wing)
         newWing.location = (0,0,0)
         newWing.rotation_euler = (0,0,0)
         objs["Wing"].append(newWing)
         newWing.parent = center  
         newWing.matrix_parent_inverse = center.matrix_world.inverted()
-    '''
+    
      
     
     if gimme():
@@ -163,7 +166,7 @@ def genWing(center):
         o.parent = center
         o.matrix_parent_inverse = center.matrix_world.inverted()
         
-    '''
+    
     if gimme():
         l, o = genPiece((0,-.7,.7))
         objs[l].append(o)
@@ -175,7 +178,7 @@ def genWing(center):
         objs[l].append(o)
         o.parent = center
         o.matrix_parent_inverse = center.matrix_world.inverted()
-    '''
+    
 
     bpy.context.scene.update()
 
@@ -187,10 +190,17 @@ bpy.context.scene.objects.link(c1)
 c2 = bpy.data.objects.new("empty", None)
 bpy.context.scene.objects.link(c2)
 
-num = 5
+num = 100
+
+
 
 os.system("rm " + write_path + "*.png")
-os.system("rm " + write_path + "mats/*")
+
+if not os.path.exists(write_path + "mats/"):
+    os.mkdir(write_path + "mats/")
+else:
+    os.system("rm " + write_path + "mats/*")
+
 
 
 for x in range(num):
@@ -199,7 +209,7 @@ for x in range(num):
 
     c1.location = (0,0,0)
     c2.location = (0,0,0)
-    '''
+    
     if gimme():
         w1 = genWing(c1)
         c1.rotation_euler = (0,0,PI/2*random.randint(-18,18)/18)
@@ -210,11 +220,11 @@ for x in range(num):
         c2.location = (3, 2 * random.randint(-1,1), .7)
 
         camera.location = (random.randint(6,11) * -1 if random.randint(0,1) < 1 else 1, random.randint(6,11) * -1 if random.randint(0,1) < 1 else 1, random.randint(12,13))
-    '''
-    #else:
-    w2 = genWing(c2)
-    c2.location = (.3 * random.randint(-1,1), .3 * random.randint(-1,1), 0)
-    c2.rotation_euler = (0,0,PI/2*random.randint(-18,18)/18)
+    
+    else:
+        w2 = genWing(c2)
+        c2.location = (.3 * random.randint(-1,1), .3 * random.randint(-1,1), 0)
+        c2.rotation_euler = (0,0,PI/2*random.randint(-18,18)/18)
 
     camera.location = (random.randint(5,7) * -1 if random.randint(0,1) < 1 else 1, random.randint(5,7) * -1 if random.randint(0,1) < 1 else 1, random.randint(6,7))
